doc: Extend the IP configuration agent API
Basically convert the flat dictionary parameter passed to the agent with
configuration bits such as local address and gateway address, into an
array of such dictionaries.  While it's possible to set up multiple
subnets on one IPv4 network this is mostly meant for IPv6 where a set of
RFCs and recommendations document how multiple Router Advertisements can
be received on a network and, if some of them don't have the "managed"
bit set, we end up with multiple addresses assigned through SLAAC.  On
our own request we may also be assigned multiple addresses within each
subnet through DHCP.

There are no plans to support such setups at this time but being a
public API it needs to be future proof.